refactor: 마켓 거래 상태 enum 을 domain.vo 로 이동
상태 enum 을 VO 계층으로 옮기고 상장 상태 알림 VO 가 도메인 모델에 의존하지 않도록 원시값만 담게 한다
