+++ api/src/main/java/ksh/tryptobackend/marketdata/adapter/in/dto/MarketStatusChangedMessage.java
@@ -1,5 +1,5 @@
 package ksh.tryptobackend.marketdata.adapter.in.dto;
 
-import ksh.tryptobackend.marketdata.domain.model.MarketStatus;
+import ksh.tryptobackend.marketdata.domain.vo.MarketStatus;
 
 public record MarketStatusChangedMessage(String exchange, String symbol, String displayName, MarketStatus status) {}
