+++ api/src/main/java/ksh/tryptobackend/marketdata/application/port/in/dto/command/ApplyMarketStatusChangeCommand.java
@@ -1,6 +1,6 @@
 package ksh.tryptobackend.marketdata.application.port.in.dto.command;
 
-import ksh.tryptobackend.marketdata.domain.model.MarketStatus;
+import ksh.tryptobackend.marketdata.domain.vo.MarketStatus;
 
 public record ApplyMarketStatusChangeCommand(
         String exchange, String baseSymbol, String displayName, MarketStatus status) {}
