+++ api/src/main/java/ksh/tryptobackend/marketdata/domain/model/ExchangeCoin.java
@@ -1,5 +1,7 @@
 package ksh.tryptobackend.marketdata.domain.model;
 
+import ksh.tryptobackend.marketdata.domain.vo.MarketStatus;
+
 public record ExchangeCoin(Long exchangeCoinId, Long exchangeId, Long coinId, String displayName, MarketStatus status) {
 
     public boolean isSuspended() {
