+++ api/src/test/java/ksh/tryptobackend/marketdata/application/service/WarmupExchangeCoinMappingServiceTest.java
@@ -17,10 +17,10 @@ import ksh.tryptobackend.marketdata.domain.model.Exchange;
 import ksh.tryptobackend.marketdata.domain.model.ExchangeCoin;
 import ksh.tryptobackend.marketdata.domain.model.ExchangeCoins;
 import ksh.tryptobackend.marketdata.domain.model.ExchangeMarketType;
-import ksh.tryptobackend.marketdata.domain.model.MarketStatus;
 import ksh.tryptobackend.marketdata.domain.vo.CoinSymbols;
 import ksh.tryptobackend.marketdata.domain.vo.ExchangeCoinMapping;
 import ksh.tryptobackend.marketdata.domain.vo.ExchangeSymbolKey;
+import ksh.tryptobackend.marketdata.domain.vo.MarketStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
